% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepHabitatModel.R
\name{clean_vri_bem}
\alias{clean_vri_bem}
\title{Title}
\usage{
clean_vri_bem(
  dsn = "./Inputs/SkWERM_Lakes.gdb",
  layer = "Lakes_10Mar2023_forVerification"
)
}
\arguments{
\item{layer}{}
}
\description{
This function imports the habitat layer (or calls it from memory), removes unnecessary columns
re-arranges the rows to allow 3 rows to represent primary, secondary and tertiary BEU
ecosystem calls and re-organizes the data table to create logical groupings for site selection
}
