% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPEM.R
\name{dom_siteseries}
\alias{dom_siteseries}
\title{Select the dominant site series to represent a PEM polygon}
\usage{
dom_siteseries(pem, wbpflag = TRUE)
}
\arguments{
\item{pem}{the pem object (from read_pem)}

\item{wbpflag}{}
}
\description{
Select the dominant site series to represent a PEM polygon
}
\details{
TO DO: fix import from or attach data.table to the package
}
