% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLayers.R
\name{get_cutblocks}
\alias{get_cutblocks}
\title{Get cutblock layer}
\usage{
get_cutblocks(tsa)
}
\arguments{
\item{tsa}{area of interest to download the consolidated cutblock layer.
Right now, just using tsa, but could replace with any area of interest}
}
\description{
Get cutblock layer
}
