% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLayers.R
\name{get_tsa}
\alias{get_tsa}
\title{Get TSA boundaries}
\usage{
get_tsa(tsa = "Lakes TSA")
}
\arguments{
\item{tsa}{which tsas. Must match exactly. To do - provide a list of valid tsa names}
}
\description{
Get TSA boundaries
}
