% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLayers.R
\name{join_ccb}
\alias{join_ccb}
\title{Join habitat layer with cutblocks}
\usage{
join_ccb(Hab_lay, aoi, gdbpath = NULL, layername = NULL)
}
\arguments{
\item{layername}{}
}
\description{
Join habitat layer with cutblocks
}
