% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLayers.R
\name{join_fire_bounds}
\alias{join_fire_bounds}
\title{Join habitat layer with Fire boundaries}
\usage{
join_fire_bounds(Hab_lay, aoi, firepath = NULL, fire_layer = NULL)
}
\arguments{
\item{fire_layer}{description}
}
\description{
Join habitat layer with Fire boundaries
}
