% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLayers.R
\name{join_FN_bounds}
\alias{join_FN_bounds}
\title{Join habitat layer with First Nations boundaries}
\usage{
join_FN_bounds(Hab_lay, bounds_path, bounds, FN_names)
}
\arguments{
\item{layername}{}
}
\description{
Join habitat layer with First Nations boundaries
}
