% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLayers.R
\name{join_huckleberry}
\alias{join_huckleberry}
\title{Title}
\usage{
join_huckleberry(Hab_lay, huck_path, huck_layername, poly_function = "mean")
}
\arguments{
\item{huck_layername}{}
}
\description{
Title
}
