% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLayers.R
\name{join_landTypes}
\alias{join_landTypes}
\title{Join with land ownership}
\usage{
join_landTypes(
  Hab_lay,
  landgdb = "D:/Github/Moose/SSAF_habitat_site_selection/data/SSAF_SiteSelection.gdb",
  landlayers = c("land_ownership_icf", "parks_ssaf_211119", "tenures_ssaf_211125",
    "woodlot_ComFor_ssaf_211126", "tfl_ssaf_211126"),
  largest = TRUE
)
}
\arguments{
\item{Hab_lay}{Habitat layer (sf) typically created in created by clean_vri_bem}

\item{landgdb}{Geodatabase pathway that contains the land ownerships.}

\item{landlayers}{names of the land ownship layers. This currently
defaults to the layers given in 2022}
}
\description{
Old function - I will delete
}
