% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLayers.R
\name{join_pem}
\alias{join_pem}
\title{Join PEM with habitat layer}
\usage{
join_pem(Hab_lay, pem_path, pem_name, largest = TRUE, wbpflag = TRUE)
}
\arguments{
\item{layername}{description}
}
\description{
This data comes as a shapefile, so no area of interest is required to define
}
\details{
This function joins the pem data to the cleaned vri-bem habitat layer. The pem layer has complex
and different delineation from the vri-bem layer. As a result, we could cut the vri-bem polygons that overlap
multiple predicted site series. However that would create new poylgons for validation, and because the
objective is to validate the already created vri-bem model, modifying the polygons for site series does
not make sense. We can  assign all site series values that overlap the vri-bem polygon (largest=FALSE),
but not provide any spatial information about where within the polygon those site series occur,
or we can only include the predicted site series for the largest overlapping area (largest = TRUE).
Currently the default is set to largest = TRUE, which means only the most prominent site series is included
but this also comes at a computational cost.
}
