% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagWBP.R
\name{join_wbp}
\alias{join_wbp}
\title{Join with C.Lamb pred wbp}
\usage{
join_wbp(Hab_lay, wbp_path, wbp_layername, poly_function = "mean")
}
\arguments{
\item{poly_function}{}
}
\description{
Join with C.Lamb pred wbp
}
