% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagWBP.R
\name{pem_wbp}
\alias{pem_wbp}
\title{Identify potential whitebark pine from PEM}
\usage{
pem_wbp(beczone, subzone, siteseries)
}
\arguments{
\item{Hab_lay}{}
}
\description{
Identify potential whitebark pine from PEM
}
\details{
the Lakes TSA PEM does not identify site series in the mkp, so can't pull out which are more likely to
contain WBP from there. Can just call out the mkp in general? Please see Land Management Handbook 26
Supplement 2 for reference to the sites series in parkland ecosystems that may contain WBP (likely too
limited in the supplement)
}
