% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagWBP.R
\name{pred_wbp}
\alias{pred_wbp}
\title{Identify potential whitebark pine from predictive map}
\usage{
pred_wbp()
}
\description{
Identify potential whitebark pine from predictive map
}
\details{
Likely use Clayton Lambs and Clason et al maps here?
}
