% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLayers.R
\name{read_cutblocks}
\alias{read_cutblocks}
\title{Read cutblock layer}
\usage{
read_cutblocks(aoi, gdbpath = NULL, layername = NULL)
}
\arguments{
\item{aoi}{area of interest to download the consolidated cutblock layer.
Right now, just using tsa, but could replace with any area of interest}
}
\description{
Read cutblock layer
}
