% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLayers.R
\name{read_FN_bounds}
\alias{read_FN_bounds}
\title{Read territorial boundaries}
\usage{
read_FN_bounds(FN_bounds_path, FN_bounds_layernames, FN_names)
}
\arguments{
\item{FN_names}{description}
}
\description{
Read territorial boundaries
}
