% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPEM.R
\name{read_pem}
\alias{read_pem}
\title{Read and process PEM layer}
\usage{
read_pem(gdbpath, layername)
}
\arguments{
\item{layername}{}
}
\description{
Read and process PEM layer
}
\details{
This function reads the pem layer. Currently we assume that it's stored locally.
An option is to apply the pem_wbp function to flag whether the pem polygon may contain whitebark pine
}
