% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLayers.R
\name{read_tsa}
\alias{read_tsa}
\title{Read TSA boundaries}
\usage{
read_tsa(gdbpath = NULL, layername = NULL, tsa = "Lakes TSA")
}
\arguments{
\item{layername}{name of the shapefile contained in a geodatabase}

\item{tsa}{which tsas. Must match exactly. To do - provide a list of valid tsa names}
}
\description{
Read TSA boundaries
}
