% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLayers.R
\name{remove_privateLand}
\alias{remove_privateLand}
\title{Remove private land}
\usage{
remove_privateLand(Hab_lay, land_path, land_name)
}
\arguments{
\item{land_name}{}
}
\description{
Remove private land
}
