% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRoadDistance.R
\name{calc_rd_dist}
\alias{calc_rd_dist}
\title{Calculate distance to nearest road}
\usage{
calc_rd_dist(aoi = "Lakes TSA", Hab_lay)
}
\arguments{
\item{Hab_lay}{}
}
\description{
Calculate distance to nearest road
}
\details{
Currently, any road. TO DO: add an argument to select road surface type
}
