% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLayers.R
\name{read_vri_bem}
\alias{read_vri_bem}
\title{Read habitat model (vri_bem) layer}
\usage{
read_vri_bem(dsn, layer)
}
\arguments{
\item{layer}{}
}
\description{
it's possible to use the output of SSGM-VRI-BEM (vri_bem.shp) here, but more likely
it'll always be passed as a shapefile stored locally (or downloaded from sync)
}
